+++ lib/commanded/middleware/pipeline.ex
@@ -1,19 +1,23 @@
 defmodule Commanded.Middleware.Pipeline do
   @moduledoc """
-  Pipeline is a struct used as an argument in the callback functions of modules implementing the `Commanded.Middleware` behaviour.
+  Pipeline is a struct used as an argument in the callback functions of modules
+  implementing the `Commanded.Middleware` behaviour.
 
-  This struct must be returned by each function to be used in the next middleware based on the configured middleware chain.
+  This struct must be returned by each function to be used in the next
+  middleware based on the configured middleware chain.
 
   ## Pipeline fields
 
-    * `assigns` - shared user data as a map.
-    * `command` - the command struct being dispatched.
-    * `consistency` - the requested dispatch consistency, either: `:eventual` (default) or `:strong`
-    * `identity` - an atom specifying a field in the command containing the
-                   aggregate's identity or a one-arity function that returns
-                   an identity from the command being dispatched.
-    * `halted` - the boolean status on whether the pipeline was halted
-    * `response` - set the response to send back to the caller
+    * `assigns` - Shared user data as a map.
+    * `command` - Command struct being dispatched.
+    * `consistency` - Requested dispatch consistency, either: `:eventual`
+       (default) or `:strong`
+    * `identity` - An atom specifying a field in the command containing the
+       aggregate's identity or a one-arity function that returns
+       an identity from the command being dispatched.
+    * `identity_prefix` - An optional prefix to the aggregate's identity.
+    * `halted` - Boolean status on whether the pipeline was halted
+    * `response` - Set the response to send back to the caller
 
   """
 
@@ -22,6 +26,7 @@ defmodule Commanded.Middleware.Pipeline do
     command: nil,
     consistency: nil,
     identity: nil,
+    identity_prefix: nil,
     halted: false,
     response: nil,
   ]
@@ -29,9 +34,11 @@ defmodule Commanded.Middleware.Pipeline do
   alias Commanded.Middleware.Pipeline
 
   @doc """
-  Puts the `key` with value equal to `value` into `assigns` map
+  Puts the `key` with value equal to `value` into `assigns` map.
   """
-  def assign(%Pipeline{assigns: assigns} = pipeline, key, value) when is_atom(key) do
+  def assign(%Pipeline{assigns: assigns} = pipeline, key, value)
+    when is_atom(key)
+  do
     %Pipeline{pipeline | assigns: Map.put(assigns, key, value)}
   end
 
@@ -55,7 +62,7 @@ defmodule Commanded.Middleware.Pipeline do
   def response(%Pipeline{response: response}), do: response
 
   @doc """
-  Sets the response to be returned to the dispatch caller, unless already set
+  Sets the response to be returned to the dispatch caller, unless already set.
   """
   def respond(%Pipeline{response: nil} = pipeline, response) do
     %Pipeline{pipeline | response: response}
@@ -63,7 +70,7 @@ defmodule Commanded.Middleware.Pipeline do
   def respond(%Pipeline{} = pipeline, _response), do: pipeline
 
   @doc """
-  Executes the middleware chain
+  Executes the middleware chain.
   """
   def chain(pipeline, stage, middleware)
   def chain(%Pipeline{} = pipeline, _stage, []), do: pipeline